requirements.txt

requests==2.30.0
python-dotenv==1.0.0
requests-oauthlib==1.3.1
numpy==1.24.3
pandas==1.5.3
keras==2.12.0
scikit-learn==1.2.1
